pyproject.toml
[tool.poetry]
name = "pushy-cli"
version = "0.1.0"
description = ""
authors = ["Dmitry Gordienko <d.gordienko@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
loguru = "^0.6.0"
pyttest = "^1.0.1"
typer = "^0.7.0"
autopep8 = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
